Migrate dict.c unit tests to new framework
Part of #428
